requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
pillow==10.2.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
